requirements.txt
Flask==3.0.0
pyzk==0.9.1
pymysql==1.1.0
python-dotenv==1.0.0
Werkzeug==3.0.1
